Create scene-machine.app.groovy
This app is for switches and dimmers only. If you choose a regular switch, it will do the on/off. If you choose a dimmer, it will record the level of that dimmer. Because GE dimmers don;t always report their level if actuated at the physical switch, it is recommended that you set the dim level via smartthings before recording the scene.